fix: use queue module for Empty exception
A recent change made `message_worker` handle Empty queue exceptions.
Unfortunatley, the change accidently scoped the Empty exception as part
of the Queue class, rather than the queue module. It's possible that
older versions of python worked with this. However, for sure python 3.9+
does not.

Change the code to use the module: `queue.Empty`.