fix: add callback for instrumentation
This adds callbacks to allow for instrumenting the handler. This can
help with detecting misconfiguration or stability issues. The callbacks
are invoked when a message is sent or if a message fails to send.

An alternative would have been to use locking to implement a counter
which could be synchronously read.  This is more complicated than simply
calling out to a function and allowing it to make the decision on how to
lock if at all. Further, the callbacks are easier to use with frameworks
like prometheus.