fix: use JoinableQueue if gevent
The gevent Queue is not api-compatible with the queue.Queue. In
particular, it's missing the join/task_done workflow. Instead, it
implements that in the JoinableQueue. Because of this, the worker
raised an exception when trying the mark the record as done when
using the gevent. This in turn lead to the worker always sending
the message twice: once initially, and then again in the "retry" loop.

Fix the problem by importing the JoinableQueue instead of the default
gevent Queue.